requirements.txt
numpy
tensorflow
Pillow
Flask
opencv-python
easydict
matplotlib